refactor models.py into a multi-file package
